Compute the difference between two trees
This commit begins work on the `checkout` command, and will also serve
as an important building block in `log`, `merge` and other commands
built on top of those.

To perform a checkout as quickly as possible, and in a way that
preserves uncommitted changes in the index and workspace, we need to
figure out which files have changed between the current HEAD and the
commit that's been requested to be checked out. For example, say the
penultimate commit (@^) and the current HEAD (@) have the following
fully expanded trees:

    @^:                                   @:
    ├── README.md [45a4fb7]               ├── README.md [45a4fb7]
    ├── TODO.md [ec63514]                 ├─┬ config [9bc815e]
    ├─┬ config [9bc815e]                  │ └─┬ database [39995f4]
    │ └─┬ database [39995f4]              │   ├── development.json [d00491f]
    │   ├── development.json [d00491f]    │   └── production.json [0cfbf08]
    │   └── production.json [0cfbf08]     └─┬ lib [8be4dc3]
    └─┬ lib [bfe70e0]                       ├── app.rb [64bb6b7]
      └── app.rb [00750ed]                  └─┬ models [0e0262f]
                                              ├── repository.rb [7ed6ff8]
                                              └── user.rb [b8626c4]

Examining the top-level of each tree, we can see the following:

- `README.md` appears in the left and right trees with the same object
  ID, and so is unchanged between these commits

- `TODO.md` appears on the left but not on the right, so it has been
  deleted in the HEAD commit

- The `config` entry has the same object ID in both trees, and so
  everything inside that directory is identical in both versions

- The `lib` entry differs between the two trees, and so we recurse into
  it, taking `lib/` as the prefix of the paths we find.

    @^:lib                                @:lib
    └── app.rb [00750ed]                  ├── app.rb [64bb6b7]
                                          └─┬ models [0e0262f]
                                            ├── repository.rb [7ed6ff8]
                                            └── user.rb [b8626c4]

- Now we see that `lib/app.rb` has been modified, and the entire
  `lib/models` directory has been added. We recurse into `lib/models/`
  and mark every file inside it as added.

The `Database::TreeDiff` class performs this computation and returns the
result as a hash that maps `Pathname` objects to pairs of
`Database::Entry` values, representing the entry at that path in the
left and right commits, if it has changed:

    { Pathname => [Database::Entry, Database::Entry] }

An `Entry` has `oid` and `mode` fields, and either entry in the pair can
be `nil`, but not both.

    Entry = Struct.new(:oid, :mode)

In the above example, the tree diff would be:

    {
      TODO.md                  => [ec63514, nil    ],
      lib/app.rb               => [00750ed, 64bb6b7],
      lib/models/repository.rb => [nil,     7ed6ff8],
      lib/models/user.rb       => [nil,     b8626c4]
    }

This structure allows Jit commands to quickly find out which files
differ between two commits, without exploring identical directories, and
without reading every single file. Entries are compared by their object
IDs, so no blobs are read during this process.