Update the index and HEAD on checkout
In the previous commit we began the `checkout` command, including logic
that would update the workspace to reflect the contents of the target
commit. We now apply the same changes to the index such that it reflects
the workspace, and update `.git/HEAD` to that it points at the target
commit. It should then be the case that HEAD, the index and the
workspace are all in sync and reflect the same commit, so that the
`status` command reports no uncommitted changes.

This work is straightforward. In the `Command::Checkout` class, we load
the index for updates before applying the migration, and save it
afterward. We add a method `Index#remove` that removes the given path
from the index as well as any files nested under that path, in case it
is a directory. In `Repository::Migration`, we apply the planned changes
to the index. Finally, we use `Refs#update_head` to move the HEAD
pointer and complete the checkout operation.