Extract `Repository::Inspector` from `Repository::Status`
The `checkout` command attempts to preserve uncommitted changes in the
index and workspace when switching between commits. It does this by
working out which files differ between the current HEAD and the target,
and only updating those files, rather than replacing the entire tree.

However, sometimes the index and workspace may contain changes to files
that would be affected by the checkout. In such cases, we should detect
that the checkout would overwrite uncommitted data and refuse to go
ahead.

To do this, we need to reuse some of the logic from the `status` command
that compares the HEAD tree, index and workspace and works out what's
changed. Rather than performing a full scan of the repository, we just
want to check those files the checkout would modify. To support this,
we're extracting the following methods from `Repository::Status` into a
new class, `Repository::Inspector`:

- `trackable_file?` checks whether untracked content exists at the given
  path. This content could be a file or a directory containing untracked
  files.

- `compare_index_to_workspace` compares an index entry to a filestat
  from the workspace. It first compares them by size, mode, and
  timestamps, and finally by hashing the file. It does not update the
  timestamps in the index if the file is unchanged; that logic remains
  in `Repository::Status`.

- `compare_tree_to_index` compares an entry from a tree to one from the
  index, comparing their modes and object IDs.

`Repository::Status` now uses `Repository::Inspector` to perform these
checks, and the same checks can be reused by other commands.