Detect conflicts during checkout
Given the `Repository::Inspector` class extracted in the last commit,
`Repository::Migration` can now check whether any of the files it plans
to change contain uncommitted changes that would be lost if the checkout
went ahead. Recall that a tree diff between HEAD and the target commit
is a hash of the form:

    { Pathname => [Database::Entry, Database::Entry] }

A path is considered conflicting with the tree diff if:

- The current index entry for <path> does not match either of the
  entries in the tree diff on object ID and mode. Matching the first
  entry means the index is unchanged from HEAD, whereas matching the
  second means the index already contains the content from the target
  commit. Any other value means the index contains data that would be
  lost.

- There is a file at <path> but it does not match the corresponding
  index entry, i.e. there are unstaged changes to the file.

- There is a directory at <path> and it contains untracked files; these
  files would have to be removed to write the new content at <path>.

- There is nothing at <path> but there is untracked content at a parent
  path. Again, this would have to be removed to accommodate the target
  commit's content.

If any of these conditions hold, we record a conflict for the path. If
there are any conflicts as a result of planning the changes, the
migration does not go ahead and the conflicts are reported to the user.