Print results on successful checkout
When a checkout succeeds, we should tell the user what happened,
informing them about the commit they just left and the one that's now
checked out.

If they had a branch checked out before the checkout, then they've not
lost the reference to that commit and so we don't need to print
anything. However if HEAD was a direct pointer to a commit, the checkout
might have destroyed the only reference to it, so we print a message
telling the user what it was, e.g.

    Previous HEAD position was 59dd0c5 Initial revision of "git"

Similarly, if the user checked out a branch, we just print that branch's
name, otherwise we print information about the new commit.

If the user had a branch checked out (i.e. HEAD was a symref) before the
checkout, but this is no longer the case, we also inform them about
"detached HEAD" state. It's still possible to make commits, it's just
that no branch will refer to them, so they may be lost if another branch
is checked out.

(The `commit` command does not yet move branch pointers when HEAD is a
symref, but this will change shortly.)