Update branch pointers on commit
To complete the utility of the `branch` and `checkout` commands, we need
the `commit` command to make branch pointers track commit chains, rather
than just moving HEAD around. That is, if `.git/HEAD` contains

    ref: ref/heads/foo

then `commit` should write the new commit ID to `.git/refs/heads/foo`,
rather than overwriting `.git/HEAD` and severing the link between HEAD
and the branch.

To enable this, we make `Refs#update_head` follow symrefs until it
reaches a ref that either does not exist or contains an object ID. As
this chain of refs is followed, each ref in turn is locked so that the
ref does not change between the time it's read and the time its target
is updated. Once the ultimate target ref is updated, all locks on refs
leading to it are released.