List branches when `branch` is called without arguments
Whereas running `branch topic` should create the `topic` branch, running
`branch` by itself should list the branches that exist. If the
`--verbose` or `-v` option is used, it should print a little information
about the tip of each branch -- its commit ID and the commit's title
line.

To present this information, the `Refs::SymRef#short_name` has been
added. This returns the path to a ref, with `.git/[refs/[heads/]]`
stripped off if the ref resides in one of those directories.