Delete branches using `--delete` and/or `--force`
Here we introduce the `--delete` and `--force` options to the `branch`
command, and the `-d`, `-f` and `-D` shorthands. (`-D` is short for
`--delete --force`.)

The `--delete` option causes a branch to be deleted, and the commit the
branch was pointing at is printed. We only delete the branch pointer,
not the commit it points at, so no database content is lost. Another
branch could be created pointed at this commit as long as the user
records its ID.

The `--force` option is used in Git to force the deletion of a branch
that has not yet been merged into the current HEAD. We don't yet have
any facilities for inspecting the history to check this, but we'll
reproduce the option and do nothing unless `--force` is set.

In `Refs`, the `delete_branch` method locks the branch pointer using a
`Lockfile`. This ensures the value of the pointer does not change
between reading the file, and deleting it.