One-line log format
A common option used with `log` is `--oneline`, a shorthand for
`--pretty=oneline --abbrev-commit`. This prints a very short
representation of each commit that can be quickly scanned: a single line
showing the commit's ID and title line.

This option only abbreviates commits if `--no-abbrev-commit` is not
used; this is accomplished by setting `:abbrev` to `:auto` by default
and setting it to a concrete boolean value when an abbreviation option
is used.