Display patches in commit logs
Having extracted diff-printing logic into `Command::PrintDiff`, and set
up support for parsing the `--[no-]patch` and `-p`, `-u` and `-s`
options, we can now add support for the `--patch` option to the `log`
command. Unlike the `diff` command, `log` should not print patches by
default.

To display the patch for a commit, we compute the tree diff between its
parent and itself, which returns a hash of the form

    { Pathname => [Database::Entry, Database::Entry] }

For each path in the tree diff, we construct two `PrintDiff::Target`
objects, one from each `Database::Entry`. These targets are then passed
to `print_diff` and the existing diff-printing logic takes over.