Extract `RevList` from the `log` command
As we evolve the functionality of the `log` command, it will be useful
to extract the logic that deals with selecting and iterating over the
commits to display, leaving the display logic within `Command::Log`.
This selection/iteration logic will also be useful in future commands.

Mirroring Git's `rev-list` command which performs a similar task, we
call this class `RevList`. Currently it replicates the logic from the
`log` command: it is instantiated with the name of a commit (defaulting
to HEAD), and its `each` method yields every commit from there backwards
by following `parent` links.