Filter tree diffs by path
In the `log` command we'd like to be able to list all the commits that
modify a certain file or directory, e.g. `log Rakefile`. To do this, we
need to take the tree diff of a commit against its parent, and see if
this diff includes the paths we're interested in.

To avoid unnecessary work, it helps to apply this filtering in the
`TreeDiff` class itself -- rather than working out the full tree diff
and then filtering it, we could skip checking loads of paths we're not
interested in. This is particularly useful given that `log` will run
this check for every commit in the history.

This commit modifies `TreeDiff` so it can take a `PathFilter`, which if
present will make the class only investigate paths matching one of these
inputs. Non-matching paths are not checked further (e.g. by recursing
into directories), nor are they returned in the tree diff result.