Logs filtered to commits changing a certain file
In the previous commit we altered `TreeDiff` so that it could filter
down to specific files of interest rather than producing the full tree
diff for a commit. We now use this to produce logs for a given file.

If an input to `RevList` is the name of a file in the workspace, it's
added to the `@prune` list. A `tree_diff` method in `RevList` works out
the tree diff between two commits, passing in a `PathFilter` that
selects interesting paths. For each commit, we work out its tree diff
and mark the commit `treesame` if the diff is empty, meaning the
interesting files were not changed. Treesame commits are filtered out of
the results.