Don't display patches for merge commits
When running `log --patch`, we should not attempt to generate and
display diffs for merge commits (commits with more than one parent).
This is because, until we support conflicted merges and merging within
files, every file in the tree of a merge commit will be equal to the
version from one of its parents. The merge itself does not introduce any
changes -- the changes come from the commits leading up to the merge.