Simplify history by only following treesame parents
When we run `log <path>`, we should see a list of the commits that led
to the file being in its current state. For merge commits, this presents
an opportunity to skip work and simplify the history.

If a merge commit is treesame to one of its parents for the given path
(i.e. the merge has the same content as the parent), that means the
merge took its version of the file from that parent, rather than
combining the versions from all of its parents. That means any changes
made on other branches leading to the merge are effectively irrelevant,
because the treesame parent's version "won" and overwrote all the other
versions.

In this case, to meet the need of explaining the file's current state,
it's sufficient to only follow the treesame parent and avoid all the
others, since the treesame parent fully explains the state of the file
in the merge.

If the merge has no treesame parents, then all its parents are followed.
For non-merge commits, this preserves existing behaviour: either the
commit is treesame to its one parent, so the parent is followed, or it's
not, so the parent is followed because there are no treesame parents.