Add `<rev>^<n>` notation for the nth parent
This change makes the `Revision` class compatible with merge commits.
The `<rev>^` notation means "the first parent of <rev>", so leaning on
`Commit#parent` is fine. But we also need a notation for selecting
parents other than the first.

Here we modify the parser to allow an optional number following the `^`
symbol. This number is then used to select which parent is needed when
evaluating the revision.