Extract `merge` logic into `Merge::{Inputs,Resolve}`
To support upcoming work, it will help us to break the existing `merge`
functionality into a couple of distinct pieces.

The `Merge::Inputs` class encapsulates identifying the two input commits
(HEAD and the requested merge target), resolving revisions to IDs, and
calculating the best common ancestor. Separating this abstraction means
the merge mechanics in the `Resolve` class can be executed without
knowing how the bounding commits for the merge are determined.

The `Merge::Resolve` class actually executes the merge. For now, all
this means is it computes the tree diff between the merge base and
target, and applies it to the index and workspace. This leaves the
repository in a state where a merge commit can be written.