Store conflicted entries in the index
When a file is conflicted following a merge, we store up to three
entries for it representing its stage in the merge base commit, in the
HEAD commit, and in the merge target. These versions are passed as a
trio to `Index#add_conflict_set`, which evicts the stage-0 entry and
stores a new entry for each version, if present.

`Index#conflict?` is needed to check whether any conflicts exist in the
index -- this will prevent commits being made while in this state.

The entries created for conflicts do not have any filestat data, because
they don't reflect files in the workspace.