Detect conflicts during merge
So far our merge implementation has assumed that the HEAD branch does
not change any of the files affected by the target branch, and so it's
safe to just apply all the changes from the target. For example:

    A   B   C   D
    o---o---o---o [master]
         \
          o---o [topic]
          E   F

Say we're on `master` and run `merge topic`. The changes introduced by
`topic` are the difference between B and F, or `F - B`. Our `merge`
command just applies a migration for `F - B` to the current HEAD and
commits the result.

However, if `master` has changed some of the same files as `topic`, that
is `D - B` and `F - B` overlap, we may have a merge conflict. We need to
decide which version from each side of the merge to keep, or enter an
error state if we can't decide. Therefore, we now work out the tree diff
of both sides of the merge and compare them to detect these kinds of
conflicts:

- both sides of the merge have modified a file present in B
- one side has modified the file and the other has deleted it
- both sides have added a file at the same path with different content

While detecting these conflicts, we attempt to merge the object ID and
mode of each entry where both sides modify the same path. If this is
possible then the conflict is automatically resolved, otherwise the
contents of both versions is dumped into the workspace and a conflict is
written to the index.

This process results in a "clean" diff being constructed that will go
through the migration machinery without problems, as the pre-image for
each entry matches the state of HEAD, assuming no uncommitted changes.

If the index is left in conflict following this operation, the `merge`
command exits unsuccessfully.