Display conflicts in the status
In order for the user to fix up merge conflicts, they need to be able to
check on the state of the repository and find out which files are still
in conflict. For this, we update the `status` command to detect and
report conflicts.

The `Repository::Status` class separates out those entries having
non-zero stages into a distinct structure, named `conflicts`. The
structure records which stages are present for each conflicted path,
allowing us to tell which type of conflict it is.

The `Command::Status` class can then use this structure to generate the
right output; the class has UI copy for every possible set of stages and
it just needs to pick the right one from the stages present in the
conflict data.