Resolve conflicts using the `add` command
The `add` command is used to indicate the current workspace version
should go into the next commit. In the case of a merge conflict this
implies the user has examined the possible versions and selected the
conflict they want to keep, constructing a resolved version.

Therefore the `add` command should cause any non-zero staged entries to
be removed for the given path, so that it is no longer conflicted.