Resume commit after a conflicted merge
When we run `commit` or `merge --continue` with an in-progress merge,
Jit should attempt to create a merge commit. If there are still
conflicts in the index, this must fail. Otherwise, we retrieve the merge
state from `.git/MERGE_{HEAD,MSG}`, delete those files and write a merge
commit.

If those state files do not exist, then `commit` will create a normal
commit, and `merge --continue` will fail.

If we try to start a new merge when these state files do exist, that
should also fail.