Break combined diffs into hunks
When we display combined diffs, we should isolate the regions of change
as hunks, just as we do for normal diffs. For example, given the
following versions of a file from the base of a merge, the HEAD commit,
the target commit, and the merge commit:

      Base          Ours          Theirs        Merged
    -----------------------------------------------------------
      alfa          alfa          echo          echo
      bravo         bravo         bravo         bravo
      charlie       delta         charlie       delta
                                                foxtrot

The diff of the Merged version against Ours and Theirs respectively
would be displayed, with hunk markers, as:

                Ours                Theirs
              ---------------------------------------
                @@ -1,3 +1,4 @@     @@ -1,3 +1,4 @@
                -alfa                echo
                +echo                bravo
                 bravo              -charlie
                 delta              +delta
                +foxtrot            +foxtrot

When displaying these as a combined diff, the two diffs are aligned on
their common content in the post-image (the Merged version), and the
hunk header should display the offsets in each pre-image (the Ours and
Theirs versions) and the post-image.

    @@@ -1,3 -1,3 +1,4 @@@
    - alfa
    + echo
      bravo
     -charlie
     +delta
    ++foxtrot

The `Hunk.filter` function currently works with arrays of `Edit` values,
and we'd like to make it work on arrays of `Combined::Row` values as
well. To see how to do this, we can look at the types of the underlying
data structures. I'm going to use a Haskell-like notation here since
Ruby does not have static types. To start with, the `Diff.diff` function
takes two strings, which it breaks into lines, and then returns a list
of edits.

    Diff.diff :: String -> String -> [Edit]

The `Edit` and `Line` structures have the types:

    Line = { number :: Int, text :: String }

    Edit = { type :: Symbol, a_line :: Line, b_line :: Line }

`a_line` is a `Line` from the pre-image, and `b_line` from the
post-image. The notation `[t]` means a list of values of type `t`. In
the `Edit` structure, either `a_line` or `b_line` can be null, but not
both.

The `Diff.combined` function takes a list of pre-images and a single
post-image, and returns a list of `Row` values, where:

    Diff.combined :: [String] -> String -> [Row]

    Row = { edits :: [Edit] }

It does this by computing a diff between each pre-image and the single
post-image, and then combining the resulting lists of `Edit` values. For
example, calling `Diff.combined([a, b], c)` combines the diffs `diff(a,
c)` and `diff(b, c)`. Each `Row` contains N edits, where N is the number
of pre-images. All the edits in a `Row` have the same `b_line`, by
definition. This means a `Row` either contains:

- a null `b_line`, meaning the edits in the row are all either deletions
  or null; it will usually contain a single deletion and all the other
  elements of `edits` will be null.

- a non-null `b_line`, meaning all the row's edits are either insertions
  or unchanged lines; in this case it's possible that all the edits are
  unchanged lines, not insertions.

For example, the first two rows in the above combined diff are `- alfa`
and `+ echo`, which are represented by rows containing the following
edits:

    [
      Edit(type = :del, a_line = Line(1, "alfa"), b_line = nil),
      nil
    ]

    [
      Edit(type = :ins, a_line = nil,             b_line = Line(1, "echo")),
      Edit(type = :eql, a_line = Line(1, "echo"), b_line = Line(1, "echo"))
    ]

Now, the existing `Hunk.filter` function takes a list of edits and
returns a list of hunks, that is:

    Hunk.filter :: [Edit] -> [Hunk]

    Hunk = { a_start :: Int, b_start :: Int, edits :: [Edit] }

We'd like to amend `Hunk.filter` so it can take a list of `Row` values,
and return hunks containing the same, that is:

    Hunk.filter :: [Row] -> [Hunk]

    Hunk = { a_start :: Int, b_start :: Int, edits :: [Row] }

The question is, can we abstract over the `Edit` and `Row` types to
produce something with a consistent interface, so that `Hunk.filter`
works with both types?

    Hunk.filter :: [t] -> [Hunk t]

    Hunk t = { a_start :: Int, b_start :: Int, edits :: [t] }

Let's remind ourselves of those types:

    Line = { number :: Int, text :: String }

    Edit = { type :: Symbol, a_line :: Line, b_line :: Line }

    Row  = { edits :: [Edit] }

`Hunk.filter` relies on the fact that the items in the list respond to
`type`, so it can determine regions of change that should be chunked
together. The functions for printing diffs also want this information to
decide which colour to use. `Hunk.filter` also relies on the `a_line`
and `b_line` properties of edits, and the `number` of those lines, to
calculate the header offsets. So in short, `Hunk.filter` relies on this
interface, which is implemented by `Edit`:

    type   :: t -> Symbol
    a_line :: t -> Line
    b_line :: t -> Line

Notice this difference between the normal and combined hunk headers:

    normal:     @@ -1,3 +1,4 @@

    combined:   @@@ -1,3 -1,3 +1,4 @@@

A normal hunk header has only one `-` offset, which comes from the
`a_line` of its edits. Combined hunks have multiple `-` offsets -- one
from each `a_line` in the edits of the combined rows. Now, having a
single `a_line` is just a special case of having many `a_lines`; a
normal diff against a single pre-image is a special case of a combined
diff against multiple pre-images. So we can adjust our interface like
so:

    type    :: t -> Symbol
    a_lines :: t -> [Line]
    b_line  :: t -> Line

`type` remains singular, because we want each `Row` to be categorised as
a single type for the purposes of hunk filtering and colour-formatting
the diff. `b_line` remains singular because by definition all the edits
in a row have the same `b_line`; only their `a_line` values differ.

We're left with the question of whether `a_lines` can be defined for
edits, and whether the entire interface can be defined for rows. The
first is straightforward: `a_lines` for an `Edit` is just a list
containing its `a_line`.

    def a_lines
      [a_line]
    end

Defining the other functions for `Row` is also fairly simple. Recall
that a `Row` either contains deletions and null entries, or insertions
and/or unchanged lines. That means `type` can be defined for `Row` to
return `:ins` if the row contains any insertions, and the `type` of the
first non-null `Edit` otherwise. This means that if there are any
deletions or insertions present then the `Row` will be classed as such,
otherwise it will be considered an unchanged line.

    def type
      types = edits.compact.map(&:type)
      types.include?(:ins) ? :ins : types.first
    end

The `a_lines` for a `Row` are just the `a_line` for each of its edits,
allowing for the fact the edits may be null:

    def a_lines
      edits.map { |edit| edit&.a_line }
    end

And the `b_line` of a `Row` is the `b_line` of the first `Edit`, since
all the edits in a `Row` have the same `b_line`. This is true even if
the first `Edit` is null, because it will be part of a `Row` containing
only deletions, which have no `b_line`.

    def b_line
      edits.first&.b_line
    end

This commit adds the necessary methods to `Diff::Edit` and
`Diff::Combined::Row` to accomplish this abstraction, and adjusts the
`Diff::Hunk` class so that it can work with an arbitrary number of
pre-images. This continues to work for normal diffs since they're a
special case of this more general behaviour.