Print combined diffs for conflicted files
Having defined `Diff::Hunk.filter` over combined diffs in a recent
commit, we can now make `diff` behave correctly when there are
conflicted entries in the index. If there are both stage-2 and stage-3
entries, and the user has not specified a stage via the command-line
options, then we print a combined diff.