Add the `log --cc` option for showing patches for merge commits
Having defined `Diff::Hunk.filter` over combined diffs in a recent
commit, we can now add the `--cc` option to `log`, which prints patches
for merge commits. We select the paths for which the merge commit
differs from all of its parents, and generate a combined diff for each
one.